Add collision support to map importer
This first imports all collisions which share the current map section prefix, using the same folder where dff files are expected. Subsequent imports with the same section prefix will ignore already loaded collisions. All collisions are placed within a collection named for the .col file from which they came, and these collections are set to hidden so as not to obscure the map geometry. Then when importing the map geometry, each collision file is instanced, transformed, and linked to the collection of the map instance, and initially set to hidden again so as to not obscure the map geometry. This change enables easy in-place editing of an instance's collision geometry.