Add .col export item to collection context menu in Outliner
This adds a menu item and some minor implementation logic to allow exporting a collection from the Outliner as a collision container (.col) file. This change complements the map collision support as it allows changes to a .col file to be exported with the original neighboring collisions in the same container. Right-click on a collection (ie. lae2_1.col) in the Outliner to export its collisions in one file which could then be used to replace the original collision container in a mod.