Replace naive collision loading with full lookup
This replaces the previous method of collision loading by region prefix with a more accurate, comprehensive lookup off all instances, loading all .col files referenced by the objects in the .ide file. This change resulted in a number of performance considerations. Firstly, the .ide list is pruned (only for GTA:SA) to exclude IDE files obviously unrelated to the current IPL (might want to revisit this). Next, the originating .ide filename is inserted in the instance data so the correct .col file can later be found (because the .col file sets are named after the .ide of the instance). This changes a number of data structures in map_data. Finally, the timer-based loading had to be adjusted to favor loading larger batches of objects at a time. Unfortunately as the number of objects in the scene increases, the implicit scene updates that Blender performs begin to crush responsiveness and prolong loading considerably. Since we can't control Blender's scene updates the only option is to load more objects each timer pulse. This makes Blender appear to freeze but is significantly faster overall.