Face group support for (.col) import/export
This implements face groups for .col files by creating a face attribute on the collision mesh in Blender and storing the face group index there on import. For export, this attribute is looked for and if found all face vertices with the same index are used to build min/max bounding boxes which are then stored along with the beginning/ending face index as a face group block in the resulting .col file. There isn't yet a way to modify these groups or generate them inside Blender although that would be something to look into. This support is currently just so a subsequent export of a collision collection will maintain any face groups that were imported so this information isn't destroyed. However, if a collision mesh is modified in some way then this attribute should be deleted for that object before export as its face groups would obviously no longer be valid.