Make .col file import to collection
Small change to move collision items to a collection named for the container file they are imported from. This is more consistent with how map importer works, it better organizes the scene, and it makes it simpler to re-export back to the file from the context menu of the outliner. Also a small gui change to state that the importer works with .col as well as .dff files