Add 'empty' collisions to import/export
This small change handles collisions which have no shapes/meshes but still define a bounds. Some .col files contain such collisions (ie. pigpen_props inside lae2_5.col). I'm actually not sure on whether these are even used by the game but this is an effort to keep as much of the original .col file intact on export as possible. The bounds are just stored in a custom property of the particular col's collection on import, and fetched from there on export.

There is also a small fix that prevents a script error when trying to export a mesh with face groups to a version 1 .col file which doesn't support that.  Which is probably an unintentional action, but script errors are still bad.