Ignore map collisions for DFF
Small change to ignore map collisions on DFF export. Map collisions are meant as a convenient way to edit the collisions of map objects in-place and then export them later as separate COL files. The way they're imported in the same collection as map geometry makes them get picked up by the DFF exporter which is likely unintended, so this just ignores them for DFF exports. It should not affect regular embedded collisions for DFF files because those are imported into their own nested collection.