Added Outliner context menu function to export singular DFFs
Firstly, a small change was made to how collections are considered during an export so that the now more nested structure of collections created by the map importer can work with DFF exports.
Next, a shortcut was added to the Outliner context menu to allow for exporting just the active object as a .dff file. This would be most useful for making changes to a map object and wanting to quickly export just that object with good default settings (correct filename, no collision, position reset to origin).
Also renamed the previous collision collection shortcut func for consistency.