Small change to speedup first map region imports
This just speeds up the initial region map import to be in line with what it was before this PR. Originally, 10 map objects are loaded at a time, but this needs to be increased as the overall scene object count grows due to background scene updates taking longer and longer. Still, at least 10 objects need to be loaded at a time when the object count is low, otherwise it takes like ~6x longer for the first import.