Small fixes to dff exporter
Fixed a regression where meshes were skipped for export unless they resided in a collection named with ".dff" at the end. Also made the exporter ensure the correct file extension when doing a 'mass export' (files are named after their collection with that option enabled, but the collection may not have ".dff" at the end of its name). Also fixed a small bug where exporting with 'export collisions' enabled will throw an error if no collisions exist in the scene.